Fix exception and add checkmarks to see which operation mode you are in
Goal of this commit was to fix a bug that occured when the input in the
input fields was empty (and with that invalid) and to address the problem
that the current operation was not visible when clicked which was solved by
putting in checkmarks that are not clickable for the user but get changed
with each iteration of an operation change

Fixed an exception that occured when the float on input_1 or input_2 was invalid (ll. 151 & 157)
Added boolean to check if any operations were selected before (l. 52)
When operation is selected checkmark gets checked and others unchecked (ll. 70-140)